code/reposition/requirements.txt
certifi==2019.11.28
mysql-connector==2.2.9
numpy==1.18.1
pandas==1.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
